r5164: reveret previous change

+++ get_clock.h
@@ -50,7 +50,7 @@ static inline cycles_t get_cycles()
 #elif defined(__PPC__) || defined(__PPC64__)
 /* Note: only PPC CPUs which have mftb instruction are supported. */
 /* PPC64 has mftb */
-typedef unsigned long cycles_t;
+typedef unsigned long long cycles_t;
 static inline cycles_t get_cycles()
 {
 	cycles_t ret;
